Optimized and cleaned up guide
