+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/MultiTool.java
@@ -26,7 +26,7 @@ public class MultiTool extends DamagableChargableItem {
 	public void postRegister() {
 		List<Integer> list = new ArrayList<>();
 		for (int i = 0; i < 50; i++) {
-			if (Slimefun.getItemValue(this.getID(), "mode." + i + ".enabled") != null && (Boolean) Slimefun.getItemValue(this.getID(), "mode." + i + ".enabled")) list.add(i);
+			if (Slimefun.getItemValue(this.getID(), "mode." + i + ".enabled") != null && (boolean) Slimefun.getItemValue(this.getID(), "mode." + i + ".enabled")) list.add(i);
 		}
 		this.modes = list;
 	}
