+++ src/main/java/me/mrCookieSlime/Slimefun/guides/SlimefunGuideLayout.java
@@ -1,4 +1,4 @@
-package me.mrCookieSlime.Slimefun.api;
+package me.mrCookieSlime.Slimefun.guides;
 
 public enum SlimefunGuideLayout {
